requirements.txt
freetype-py
PyOpenGL
PyOpenGL_accelerate